Refactor import statement for TaskDetails component
